Option to lock any kind of setup item.
