+++ main/setup/SetupMenuChar.h
@@ -9,7 +9,6 @@
 
 #include "setup/SetupNG.h"
 #include "setup/MenuEntry.h"
-#include "setup/SetupMenuValCommon.h"
 
 class SetupMenuChar:  public MenuEntry
 {
@@ -38,7 +37,6 @@ private:
 	mutable int  _select = 0;
 	int  _select_save = 0;
 	int  _char_index = 0;   // position of character to be altered
-	bitfield_select bits = {};
 	std::vector<const char *> _values;
 	int (*_action)( SetupMenuChar *p );
 	char *_mychar = nullptr;
