+++ main/setup/SetupMenuCommon.h
@@ -1,0 +1,3 @@
+#pragma once
+
+typedef enum { RST_NONE, RST_ON_EXIT, RST_IMMEDIATE } e_restart_mode_t;
